refactor: use existing get recipe file method
